Rooms now open doors upon generation
During generation, the rooms cast a sphere on each of their four doors to see if the sphere intersects with another room collider that is not their own. If it does intersect, then the room will open the door that connects with that other adjacent room.

Coroutines really make the whole process more performant, because it doesn't try to do everything at once.