Minimap Implemented, Player Camera Tweaked
So, the first part - the minimap.

The minimap is a little rendure texture on the right hand corner of the player's HUD that tells the player his current location in the whole map. This is done by using an orthographic camera somewhere far off in the scene that captures the whole map, and the player is represented by a giant sphere that is invisible to the player, but not to the orthographic camera.

The Player camera also got a tweak, that is, weapons will now always show in front, preventing it from 'clipping' into walls.