Ammo System Added
Guns now require ammo to fire. The amount of ammo each gun consumes is dependent on the ammoConsumption variable set in the GunSettings class. If the current amount of ammo (which is in the form of an IntVariable scriptable object) is not sufficient to fulfill the ammoConsumption, then the gun won't fire.