Ammo Drops drop from enemies
Whenever an enemy is killed, there is a chance that it would drop an ammo drop game object into the scene by instantiation.

The gameobject prefabt that is being instantiated is stored in a scriptable object. In that scriptable object, there is a function that returns the prefab, if it passes a random dice roll of sorts.