Room Generation Sets Added
Room Generation Sets are asset files that contain the data that the Map Generator Script uses to decide which rooms to generate.

Inside this Room Generation Set is a serializable class that contains the gameobject room prefab and the probability. In this same scriptable object script file, there is a function that makes use of cumulative probability (I think) to return the game object prefab stored in that serializable class according to its assigned probability variable.