Basic weapon modules system structure added
For now, the structure of the weapon modules system isn't too sophisticated.

There are abstract events in the abstract base class so that they can be re-initialized in the child class by force. This then allows scriptable objects containing abstract functions to subscribe to these events. As these are abstract functions from an abstract scriptable object class, these abstract functions can be filled in by a wide variety of code by its child classes.

Therefore, the functions contained within a class that inherits from the weapon module class can be invoked via event in the child classes of the weapon class.