Weapon Remove bug fixed
So I learnt an important lesson today.

When executing functions inside of a foreach loop, make sure that that function does not modify the list that the foreach loop is iterating through. Otherwise, you get that weird error in Unity saying that a collection was modified and it could not proceed.

To solve this, I made a separate function for removing modules from the list. This is different from the function that removes the effects of a module on a weapon. Then, I made the player weapon controls script call those functions separately not in a foreach loop. That way, I'm not iterating through any loops, and just remove the modules individually.