Added Maximum Module Count
So, sort of unrelated, but in order to make the display work, I need to know how many modules there can be at any given moment. So, I added an int variable and a little logic in the ApplyModule method in the Gun script to make it so that no more modules would be added if the module count is already at maximum as defined in the int variable scriptable object.

Oh, and the display works nicely now.