Imported the Unity NavMesh Components Scripts
So, in order to be able to bake NavMeshes in Runtime, I had to import a set of scripts from Unity's GitHub repository. Yesterday, I ran into a lot of errors from the console, which said that quite a few scripts were missing, although they were clearly there in the Assets folder.

Anyway, I finally figured out today that if I just use the Import New Assets options and then import the scripts that way, I got past the error, and the scripts seem to be working just fine.