Test Rooms have their own NavMeshSurfaces
I've added NavMeshSurface components to each of the three test battle rooms I have right now.

This means that every room would have its own NavMesh that is specific to only that room, which is exactly what I want. The best part is, it's already baked into the room prefab itself, so I don't have to run a function to bake the NavMesh on the room prefab themselves on Awake.