Basic Melee Enemy Behavior Added
Finally, after all this time, I have an enemy that uses the NavMesh navigation system in order to track down the player, and then stop at a certain distance from the player.

Enemy behavior consists of two main components: the BaseEnemyBehavior component and the complimentary component (not actually named 'complimentary'), which is the component which would use the basic function of the base enemy behavior component (running a state machine based on the player's distance) to execute its own functions within the state machine in the base component.