Projectile-Throwing Enemy Added
I made a projectile-throwing enemy in the game.

This projectile enemy can lob projectiles in an arch while far away from the player, and throws projectiles quickly in a straight line when near the player.

I made this enemy behavior using the base enemy behavior.

To make this possible, there are several things that I made to support this feature.
1. A projectile launcher script that is general-purpose, meaning that it can be used by other things other than enemies.
2. A projectile behavior script that supports other scripts that subscribe to the events in this script. This is used by the projectile that is being launched.