Grenades Added
So these grenades are really just a modified version of the projectile.

I made the explosion script in the same way I made the damage projectile script.

The way that the script works is that it starts a timer when it touches the ground. Then after that timer goes off, it produces a sphere cast at the position where the 'grenade' is at, and caches that into a local Collider array.

Next, the that array is iterated through using a for loop, instead of a foreach loop, because I learnt my lesson. For every element, grab the Health component using the GetComponent function, and then deal damage according to how far away the object is from ground zero of the grenade.

The thing that needs to be worked on next, is obstacle avoidance. I'm thinking this can be done by raycasting towards each of the objects in the Collider array.