Explosive Projectiles Detect Obstacles
Now with the addition of a raycast test to determine whether an object within a sphere cast is unobstructed from any objects within the scene, the explosive projectiles will not be able to damage targets through a wall for example.

More specifically, the way that this test works is that after sphere casting and collecting all relevant colliders (target colliders), a ray will be cast in the direction of each of the target colliders. If it returns false, then the explosion is unobstructed and that particular collider's Health component can then be damaged.