Weapon Module System Redesign
After trying to make a weapon module that would fire a grenade after every five shots, I realized that the previous method of making and applying weapon modules was way too overcomplicated.

I came up with a new system where every component on a weapon is its own system, which was capable of storing and accessing data needed for its function at runtime.

For now, I made the 'base' module that most weapons would have - the RaycastShooter module, which like the name suggests, shoots out a raycast when the player presses on the left mouse button.