Full Auto Module Added
I added the full auto mod to the assault rifle as a sort of demo to prove that this new system of weapon modules allows each module to have its own functionality and dependencies.

In the case of this full auto module, it can store the time since its last attack within the assault rifle gameobject. It can also reference the base module - the raycast shooter, to do the shooting for the full auto module.

Let's just see where this design would lead to.