Module activation added
So I made the 'WeaponType' scriptable object, which just contains a list of 'WeaponComponentType' scriptable objects, which would be compatible with that type of weapon.

Using this, I made a monobehavior called a WeaponComponentActivator, which as its name suggests, activates components, but only if a method inside of it is called by some other script from outside the gameobject. This is where the 'ModuleGiver' comes in.

Yeah, yeah, I know that 'WeaponModules' and 'WeaponComponents' have different names, but I still see them as the same thing.