Pistol now shoots ammo packets
I was thinking of ways to give the player ammo, without substantially slowing down the combat of the game.

I came up with the idea of shooting ammo packets from the pistol, which is a weapon that not many people even think about in most shooters, to give the pistol a reason to exist - as an ammo generator that doubles as a good enough weapon, until the player has enough ammunition to use their main weapons again.