Players can now save and quit
Using the level lists and player prefs, I was able to pretty quickly to implement the feature, where in the pause screen (making the pause screen is nothing special), the player can click on a button that would save the last left off level into PlayerPrefs, and send them off into the title screen.

Also, on the main title screen, there is an option to continue the last left off play session. There's a dedicated script for this one button, called ContinueButton. This script checks whether or not there's a level saved (in the form of an integer of the build index), and if there is, then it would turn the button interactable, if not then non-interactable.