Sensitivity adjustments in option screen
Added an options menu, and for now, the only thing on there are two sliders for mouse sensitivity.

This was done by making use of PlayerPrefs, and making a new type of scriptable object - PlayerSettings. PlayerSettings just stores all the other scriptable object variables needed to store information about the Player's Settings. These settings are loaded from PlayerPrefs on Awake in the Title Screen by a small script, called SettingsLoader.