Module Stands Added
A module stand is an object that appears by chance at the end of every encounter.

I made this by writing the WeaponComponentStand script, which would execute the method that rolls for a weapon component type in the WeaponComponentReward. All the probabilities and what not are all stored in the WeaponComponentRewardSet.
If the function returns null, then the stand deactivates itself, if not, it activates. When it activates, it chooses a random weapon component type to give.