Some Small Improvements Made
Well, technically all the main features of the game are already there:
a first person player,
some enemies that can walk through a navmesh,
level generation,
a speedrun clock,
and finally a way to progress through levels to the victory screen.

From here on out, it's mostly just adding in the 'flesh', such as making a bunch of different rooms with a bunch of different enemies.